The parser successfully parses tic-tac-toe
I decided to scrap the python proof of concept.  While I do code faster
in python than anything else.  The time saved would not make up for the
time wasted getting it to work when I'm just going to toss it out in
favor of Flex/Bison.

Next step will be to generate and AST and then the C code after that.